+++ craft_providers/lxd/lxd_instance.py
@@ -17,30 +17,40 @@
 
 """LXD Instance Executor."""
 
-import hashlib
 import io
 import logging
 import os
 import pathlib
-import re
 import shutil
 import subprocess
 import tempfile
+import warnings
 from typing import Any, Dict, Iterable, List, Optional
+
+import yaml
 
 from craft_providers import pro
 from craft_providers.const import TIMEOUT_SIMPLE
 from craft_providers.errors import details_from_called_process_error
-from craft_providers.executor import Executor
+from craft_providers.executor import Executor, get_instance_name
 from craft_providers.lxd.errors import LXDError
 from craft_providers.lxd.lxc import LXC
 from craft_providers.util import env_cmd
 
 logger = logging.getLogger(__name__)
 
+PRO_SERVICES_YAML = pathlib.Path("/root/pro-services.yaml")
+
 
 class LXDInstance(Executor):
-    """LXD Instance Lifecycle."""
+    """Wrapper for a LXD Instance.
+
+    :ivar name: The provided name for the instance.
+    :ivar instance_name: The normalized name actually used for the instance.
+    :ivar project: The name of the LXD project.
+    :ivar remote: The name of the LXD remote.
+    :ivar lxc: The LXC wrapper to use.
+    """
 
     def __init__(
         self,
@@ -50,17 +60,21 @@ class LXDInstance(Executor):
         project: str = "default",
         remote: str = "local",
         lxc: Optional[LXC] = None,
+        intercept_mknod: bool = True,
     ) -> None:
         """Create an LXD executor.
 
         To comply with LXD naming conventions, the supplied name is converted to a
         LXD-compatible name before creating the instance.
 
-        :param name: Unique name of the lxd instance
-        :param default_command_environment: command environment
-        :param project: name of lxd project
-        :param remote: name of lxd remote
-        :param lxc: LXC instance object
+        :param name: The name of the LXDInstance.
+        :param default_command_environment: The command environment.
+        :param project: The name of the LXD project.
+        :param remote: The name of the LXD remote.
+        :param lxc: The LXC wrapper to use.
+        :param intercept_mknod: If the host can, tell LXD instance to intercept mknod
+
+        :raises LXDError: If the name is invalid.
         """
         super().__init__()
 
@@ -70,65 +84,15 @@ class LXDInstance(Executor):
             self.default_command_environment = {}
 
         self.name = name
-        self._set_instance_name()
+        self.instance_name = get_instance_name(name, LXDError)
         self.project = project
         self.remote = remote
+        self._intercept_mknod = intercept_mknod
 
         if lxc is None:
             self.lxc = LXC()
         else:
             self.lxc = lxc
-
-    def _set_instance_name(self) -> None:
-        """Convert a name to a LXD-compatible name.
-
-        LXD naming convention:
-        - between 1 and 63 characters long
-        - made up exclusively of letters, numbers, and hyphens from the ASCII table
-        - not begin with a digit or a hyphen
-        - not end with a hyphen
-
-        To create a LXD-compatible name, invalid characters are removed, the name is
-        truncated to 40 characters, then a hash is appended:
-        <truncated-name>-<hash-of-name>
-        └     1 - 40   ┘1└     20     ┘
-
-        :param name: name of instance
-        :raises LXDError: if name contains no alphanumeric characters
-        """
-        # remove anything that is not an alphanumeric characters or hyphen
-        name_with_valid_chars = re.sub(r"[^\w-]", "", self.name)
-        if not name_with_valid_chars:
-            raise LXDError(
-                brief=f"failed to create LXD instance with name {self.name!r}.",
-                details="name must contain at least one alphanumeric character",
-            )
-
-        # trim digits and hyphens from the beginning and hyphens from the end
-        trimmed_name = re.compile(r"^[0-9-]*(?P<valid_name>.*?)[-]*$").search(
-            name_with_valid_chars
-        )
-        if not trimmed_name or not trimmed_name.group("valid_name"):
-            raise LXDError(
-                brief=f"failed to create LXD instance with name {self.name!r}.",
-                details="name must contain at least one alphanumeric character",
-            )
-        valid_name = trimmed_name.group("valid_name")
-
-        # if the original name meets LXD's naming convention, then use the original name
-        if self.name == valid_name and len(self.name) <= 63:
-            instance_name = self.name
-
-        # else, continue converting the name
-        else:
-            # truncate to 40 characters
-            truncated_name = valid_name[:40]
-            # hash the entire name, not the truncated name
-            hashed_name = hashlib.sha1(self.name.encode()).hexdigest()[:20]
-            instance_name = f"{truncated_name}-{hashed_name}"
-
-        self.instance_name = instance_name
-        logger.debug("Set LXD instance name to %r", instance_name)
 
     def _finalize_lxc_command(
         self,
@@ -408,8 +372,14 @@ class LXDInstance(Executor):
                 uid = os.getuid()
             config_keys["raw.idmap"] = f"both {uid!s} 0"
 
-        if self._host_supports_mknod():
-            config_keys["security.syscalls.intercept.mknod"] = "true"
+        if self._intercept_mknod:
+            if not self._host_supports_mknod():
+                warnings.warn(
+                    "Application configured to intercept guest mknod calls, "
+                    "but the host OS does not support intercepting mknod."
+                )
+            else:
+                config_keys["security.syscalls.intercept.mknod"] = "true"
 
         self.lxc.launch(
             config_keys=config_keys,
@@ -684,6 +654,35 @@ class LXDInstance(Executor):
             project=self.project,
             remote=self.remote,
         )
+
+    @property
+    def pro_services(self) -> Optional[set[str]]:
+        """Get the Pro services enabled on the instance."""
+        # first check if the services are cached in memory
+        if hasattr(self, "_pro_services"):
+            return self._pro_services
+        # then check the instance state
+        try:
+            with self.edit_file(
+                source=PRO_SERVICES_YAML,
+            ) as temp_state_path:
+                with temp_state_path.open("r") as fh:
+                    return yaml.safe_load(fh)
+
+        except FileNotFoundError:
+            return None
+
+    @pro_services.setter
+    def pro_services(self, services: set[str]) -> None:
+        """Set the Pro services enabled on the instance."""
+        self._pro_services = services  # cache the services in memory ...
+        # ... and write them to the instance
+        with self.edit_file(
+            source=PRO_SERVICES_YAML,
+            pull_file=False,
+        ) as temp_state_path:
+            with temp_state_path.open("w") as fh:
+                yaml.safe_dump(set(services), fh)
 
     def install_pro_client(self) -> None:
         """Install Ubuntu Pro Client in the instance.